# football_player_service/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Football Player Service",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "summary": "Health",
        "description": "Health check endpoint.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/players": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "List Players",
        "description": "Get all players.",
        "operationId": "list_players_players_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Player"
                  },
                  "type": "array",
                  "title": "Response List Players Players Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "players"
        ],
        "summary": "Create Player",
        "description": "Create a new player.",
        "operationId": "create_player_players_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/players/{player_id}": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Read Player",
        "description": "Get a specific player by ID.",
        "operationId": "read_player_players__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "players"
        ],
        "summary": "Update Player",
        "description": "Update a player by ID (full replace).",
        "operationId": "update_player_players__player_id__put",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Player"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "players"
        ],
        "summary": "Delete Player",
        "description": "Delete a player by ID.",
        "operationId": "delete_player_players__player_id__delete",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Player": {
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 2,
            "title": "Full Name"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "title": "Country"
          },
          "status": {
            "$ref": "#/components/schemas/PlayingStatus"
          },
          "current_team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Team"
          },
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "market_value": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Market Value",
            "default": 0
          },
          "age": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Age",
            "default": 0
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "country",
          "status",
          "id"
        ],
        "title": "Player",
        "description": "Response model including server-generated ID.\n\nWhen migrating to SQLModel, we will add `table=True`\nwhile keeping the HTTP contract unchanged."
      },
      "PlayerCreate": {
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 2,
            "title": "Full Name"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "title": "Country"
          },
          "status": {
            "$ref": "#/components/schemas/PlayingStatus"
          },
          "current_team": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Team"
          },
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "market_value": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Market Value",
            "default": 0
          },
          "age": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Age",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "country",
          "status"
        ],
        "title": "PlayerCreate",
        "description": "Incoming payload with validation + normalization rules."
      },
      "PlayingStatus": {
        "type": "string",
        "enum": [
          "active",
          "retired",
          "free_agent"
        ],
        "title": "PlayingStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}